add author and h2

+++ spring-boot/src/main/java/ztw/springboot/service/interfaces/IBooksService.java
@@ -6,5 +6,6 @@ import java.util.Collection;
 
 public interface IBooksService {
     Collection<Book> getBooks();
-    Book getBook(int id);
+
+    Book getBook(long id);
 }
